Added a Variables object
